refactor: estandarización de proyecto de acuerdo a la arquitectura propuesta(#28)
