+++ src/app/components/politics/politics.component.html
