+++ src/app/components/politics/politics.component.scss
