+++ src/app/components/politics/politics.component.ts
