+++ src/app/components/test/test.component.html
@@ -1,5 +1,5 @@
 <p>test works!</p>
-<div class="carousel">
+<!-- <div class="carousel">
   <app-carousel [itemsIn]="leads"
   mode="CARD"
   [autoPlay]=true
@@ -15,13 +15,13 @@
   itemsPerPage=6
   >
   </app-carousel>
-</div>
+</div> -->
 <!--
 <div class="card">
   <app-leader-card></app-leader-card>
 </div> -->
 
-<div class="workshops">
+<!-- <div class="workshops">
   <app-side-workshop></app-side-workshop>
 </div>
 
@@ -31,4 +31,4 @@
   <app-side-communities [communities]="communities">
 
   </app-side-communities>
-</div>
+</div> -->
