+++ src/app/modules/auth/auth.module.ts
@@ -2,19 +2,19 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { MaterialModule } from '../material/material.module';
-import { FormularioAutenticacionComponent } from './formulario-autenticacion/formulario-autenticacion.component';
+
 import { SharedModule } from '../shared/shared.module';
 import { ReactiveFormsModule } from "@angular/forms";
 import { RouterModule } from '@angular/router';
 
-
+import {AuthFormComponent} from './auth-form/auth-form.component';
 import { AngularFireModule } from '@angular/fire';
 import { AngularFirestoreModule } from '@angular/fire/firestore';
 import { AngularFireAuthModule } from '@angular/fire/auth';
 import { environment } from 'src/environments/environment';
 
 @NgModule({
-  declarations: [FormularioAutenticacionComponent],
+  declarations: [AuthFormComponent],
   imports: [
     CommonModule,
     MaterialModule,
@@ -28,7 +28,7 @@ import { environment } from 'src/environments/environment';
   ],
   exports:[
 
-    FormularioAutenticacionComponent
+    AuthFormComponent
   ],
 })
-export class SeguridadModule { }
+export class AuthModule { }
