+++ src/app/modules/auth/services/auth.service.ts
@@ -3,8 +3,8 @@ import { Injectable } from '@angular/core';
 // import { credencialesUsuario, User } from 'src/app/core/models/user';
 import { environment } from 'src/environments/environment';
 import { map } from 'rxjs/operators';
-import { User, UserFirebase } from 'src/app/core/models/user';
-import { LoginUser } from 'src/app/core/models/user_login';
+import { User, UserFirebase } from 'src/app/core/models/auth/user';
+import { LoginUser } from 'src/app/core/models/auth/user_login';
 import { AngularFireAuth } from '@angular/fire/auth';
 import firebase from 'firebase/app';
 import { NotificationService } from 'src/app/core/services/notification.service';
