+++ src/app/modules/login/components/login/login.component.html
@@ -1,4 +1,4 @@
-<app-formulario-autenticacion
+<app-auth-form
 (onSubmit)="login($event)"
 [errores]="errores"
-[accion]="'login'"></app-formulario-autenticacion>
+[accion]="'login'"></app-auth-form>
