+++ src/app/modules/login/components/recovery/recovery.component.html
@@ -1,4 +1,4 @@
-<app-formulario-autenticacion
+<app-auth-form
 (onSubmit)="recovery($event)"
 [errores]="errores"
-[accion]="'recovery'"></app-formulario-autenticacion>
+[accion]="'recovery'"></app-auth-form>
