+++ src/app/modules/login/components/register/register.component.html
@@ -1,4 +1,4 @@
-<app-formulario-autenticacion
+<app-auth-form
 (onSubmit)="register($event)"
 [errores]="errores"
-[accion]="'register'"></app-formulario-autenticacion>
+[accion]="'register'"></app-auth-form>
