+++ src/app/modules/login/login.module.ts
@@ -3,7 +3,7 @@ import { CommonModule } from '@angular/common';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { MaterialModule } from '../material/material.module';
-import { SeguridadModule } from '../seguridad/seguridad.module';
+import { AuthModule } from '../auth/auth.module';
 import { LoginRoutingModule } from './login-routing.module';
 
 import { LoginComponent } from './components/login/login.component';
@@ -19,7 +19,7 @@ import { RecoverySuccessComponent } from './components/recovery-success/recovery
     LoginRoutingModule,
     MaterialModule,
     ReactiveFormsModule,
-    SeguridadModule,
+    AuthModule,
     SharedModule
   ]
 })
