+++ src/app/modules/shared/components/communities/side-communities/side-communities.component.scss
