+++ src/app/modules/shared/components/success-template/success-template.component.scss
