+++ src/app/modules/shared/mostrar-errores/mostrar-errores.component.scss
