+++ src/app/modules/shared/mostrar-errores/mostrar-errores.component.ts
@@ -3,7 +3,7 @@ import { Component, Input, OnInit } from '@angular/core';
 @Component({
   selector: 'app-mostrar-errores',
   templateUrl: './mostrar-errores.component.html',
-  styleUrls: ['./mostrar-errores.component.css']
+  styleUrls: ['./mostrar-errores.component.scss']
 })
 export class MostrarErroresComponent implements OnInit {
   @Input()
