+++ src/app/modules/shared/shared.module.ts
@@ -5,11 +5,11 @@ import { MaterialModule } from '../material/material.module';
 import { MostrarErroresComponent } from './mostrar-errores/mostrar-errores.component';
 import { CarouselComponent } from './components/carousel/carousel.component';
 import { LeaderCardComponent } from './components/leader-card/leader-card.component';
-import { SideCommunitiesComponent } from './communities/side-communities/side-communities.component';
+import { SideCommunitiesComponent } from './components/communities/side-communities/side-communities.component';
 import { RouterModule } from '@angular/router';
 import { SideWorkshopComponent } from './components/workshops/side-workshop/side-workshop.component';
 import { FormsModule } from '@angular/forms';
-import { SuccessTemplateComponent } from './success-template/success-template.component';
+import { SuccessTemplateComponent } from './components/success-template/success-template.component';
