+++ src/app/modules/ui/components/footer/footer.component.html
@@ -1,8 +1,8 @@
 <footer>
     <mat-grid-list cols="3" rowHeight="50px">
 
-        <mat-grid-tile><a routerLink="/app/terminos-condiciones">Términos y condiciones</a></mat-grid-tile>
-        <mat-grid-tile><a routerLink="/app/politicas-privacidad">Politicas de privacidad</a></mat-grid-tile>
+        <mat-grid-tile><a routerLink="/terminos-condiciones">Términos y condiciones</a></mat-grid-tile>
+        <mat-grid-tile><a routerLink="/politicas-privacidad">Politicas de privacidad</a></mat-grid-tile>
         <mat-grid-tile>
             <div class="redes-sociales">
                 <ul>
